// web/lib/api/generated/models/storageLocationResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * GroceryManager.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface StorageLocationResponse {
  id: string;
  name: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  sortOrder: number | string;
  isDefault: boolean;
  isArchived: boolean;
  version: string;
}